Linux linpack - Similar to runme_xeon64
This is the file that is very simmilar to the runme_xeon64 scripty that is aquired by the intell download. the runme_xeon64 is the linpack test but the linux.sh is simmilare to the srcipt, This is most likely what we need to change to have proper compatibility with the benchmarks we are running.